+++ mobile/android/android-components/docs/changelog.md
@@ -84,6 +84,7 @@ permalink: /changelog/
   * When binding a subscription to a `LifecycleOwner` then this subscription will automatically paused and resumed based on whether the lifecycle is in STARTED state.
   * When binding a subscription to a `View` then this subscription will be paused until the `View` gets attached.
   * Added `Store.broadcastChannel()` to observe state from a coroutine sequentially ordered.
+  * Added helpers to process states coming from a `Store` sequentially via `Fragment.consumeFrom(Store)` and `View.consumeFrom(Store)`.
 
 * **support-ktx**
   * ⚠️ **This is a breaking behavior change**: `JSONArray.mapNotNull` is now an inline function, changing the behavior of the `return` keyword within its lambda.